upload acid drip
This is an acid dripping effect for when robots are taking damage over time from acid. The last frame is designed to stay around rather than fading away, and as multiple drops fall they'll start to form puddles. Each frame is 8x8 pixels, but not every frame uses all of that space.