upload jet flame animations
jet_ignite should play once when the ship takes off, and at the end transition to jet_loop. Ideally, the frames of jet_loop would be played in random order, with a 50% chance for each one to be flipped across the vertical axis, but it would also be ok to just play each frame in order on a loop.

Each frame is 64x64 pixels.